pyproject.toml
[tool.poetry]
name = "projectjwt"
version = "0.1.0"
description = ""
authors = ["bakashev <bakashevs@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
pillow = "^9.5.0"
pyjwt = "^2.7.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
